Adjust the Windows HiDPI LAF improvements by keeping borders 1 device pixels thick. Improve JSpinner borders. Fix style.
